Move DecisionTree definition to diagnoses router
