+++ src/lib/Icons/QuestionMarkCircle.svelte
@@ -1,0 +1,1 @@
+<i class={`fa fa-question-circle ${$$props.class}`}></i>
