+++ src/routes/intake/+page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
+    import LoadingBar from "$lib/components/LoadingBar.svelte"
+    import QuestionMarkCircle from "$lib/icons/QuestionMarkCircle.svelte"
+    import { downloadBlob } from "$lib/utils"
     import { getModalStore, getToastStore, type ModalSettings } from "@skeletonlabs/skeleton"
     import { onMount } from "svelte"
-    import QuestionMarkIcon from "$lib/components/QuestionMarkIcon.svelte"
-    import LoadingBar from "$lib/components/LoadingBar.svelte"
-    import { downloadBlob } from "$lib/utils"
 
     let redcapSurveyId = ""
     let isLoading = false
@@ -72,7 +72,7 @@
 <div class="flex space-x-1">
     <label for="redcapSurveyId">MRN</label>
     <button class="hover-highlight" on:click={explainMRN} disabled={modalOpen} tabindex="-1">
-        <QuestionMarkIcon />
+        <QuestionMarkCircle />
     </button>
 </div>
 <form class="space-y-2">
